Extract num of lives and num of guesses as constants
