+++ src/components/Quiz/Player.jsx
@@ -1,7 +1,6 @@
 import { useEffect, useRef } from "react";
 
 const Player = ({ source }) => {
-  // Set default audio volume to 50%
   const audioRef = useRef(null);
 
   useEffect(() => {
@@ -14,7 +13,7 @@ const Player = ({ source }) => {
     <audio
       ref={audioRef}
       controls
-      className="m-auto opacity-50 w-[100%] hover:opacity-80 transition duration-300 ease-in-out my-2"
+      className="m-auto dark:opacity-70 w-[100%] transition duration-300 ease-in-out my-2"
       onLoadedMetadata={() => {
         if (audioRef.current) {
           audioRef.current.volume = 0.3;
