+++ src/settings.js
@@ -1,0 +1,2 @@
+export const NUM_OF_LIVES = 4;
+export const NUM_OF_GUESSES_ALLOWED = 6;
